try building the site
This is going to install and build hugo and its deps every time which is
not super desirable. Ideally I'd have a build step here that builds a
hugo container and then another one that uses it. Perhaps. We shall see.
I may end up separating out a hugo container build of my own if I don't
end up trying to port the hugo container building on github to work with
arm64 as well as amd64